feat: add mermaid rendering to FileViewer markdown preview
Extract MermaidBlock and CodeBlock from MarkdownBody into a shared component. Use them in FileViewer's TextFileViewer markdown preview so mermaid diagrams render the same way as in chat messages.

- MermaidBlock defaults to preview mode (no manual toggle needed)

- Non-mermaid code blocks use CodeBlock for consistent styling

- Streaming messages still fall back to source view via isStreaming

- 5 unit tests covering SSR initial state, streaming, and edge cases
