+++ components/MermaidBlock.tsx
@@ -1,0 +1,158 @@
+"use client";
+
+import { useEffect, useState, type ReactNode } from "react";
+import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
+import { vs } from "react-syntax-highlighter/dist/cjs/styles/prism";
+import { vscDarkPlus } from "react-syntax-highlighter/dist/cjs/styles/prism";
+import { useTheme } from "@/hooks/useTheme";
+import { copyText } from "@/lib/clipboard";
+
+interface MermaidBlockProps {
+  code: string;
+  isStreaming?: boolean;
+}
+
+/**
+ * Renders a Mermaid diagram from a fenced code block.
+ * Shows source code by default with a "Preview" toggle.
+ * When preview is active, dynamically imports mermaid and renders to SVG.
+ */
+export function MermaidBlock({ code, isStreaming }: MermaidBlockProps) {
+  const { isDark } = useTheme();
+  const [showPreview, setShowPreview] = useState(true);
+  const [svg, setSvg] = useState<string | null>(null);
+  const [renderedKey, setRenderedKey] = useState("");
+  const [failedKey, setFailedKey] = useState<string | null>(null);
+  const currentKey = `${isDark ? "dark" : "light"}\n${code}`;
+
+  useEffect(() => {
+    if (!showPreview || isStreaming) return;
+
+    let cancelled = false;
+    setFailedKey(null);
+
+    const render = async () => {
+      const { default: mermaid } = await import("mermaid");
+      mermaid.initialize({
+        startOnLoad: false,
+        securityLevel: "strict",
+        suppressErrorRendering: true,
+        theme: isDark ? "dark" : "default",
+      });
+
+      const parsed = await mermaid.parse(code, { suppressErrors: true });
+      if (!parsed) throw new Error("Invalid Mermaid diagram");
+
+      const id =
+        typeof crypto !== "undefined" && "randomUUID" in crypto
+          ? `mermaid-${crypto.randomUUID()}`
+          : `mermaid-${Date.now()}-${Math.random().toString(36).slice(2)}`;
+      const result = await mermaid.render(id, code);
+      if (!cancelled) {
+        setSvg(result.svg);
+        setRenderedKey(currentKey);
+      }
+    };
+
+    render().catch(() => {
+      if (!cancelled) setFailedKey(currentKey);
+    });
+
+    return () => {
+      cancelled = true;
+    };
+  }, [code, isDark, isStreaming, showPreview]);
+
+  const previewButton = (
+    <button
+      onClick={() => setShowPreview((v) => !v)}
+      disabled={isStreaming}
+      title={isStreaming ? "Preview available after streaming" : (showPreview ? "Show Mermaid source" : "Preview Mermaid diagram")}
+      className={["markdown-code-action", showPreview ? "is-active" : ""].filter(Boolean).join(" ")}
+    >
+      {showPreview ? "Source" : "Preview"}
+    </button>
+  );
+
+  if (!showPreview || isStreaming) {
+    return <CodeBlock code={code} lang="mermaid" headerAction={previewButton} />;
+  }
+
+  const body =
+    failedKey === currentKey ? (
+      <div className="mermaid-block mermaid-block-error">Invalid Mermaid diagram</div>
+    ) : !svg || renderedKey !== currentKey ? (
+      <div className="mermaid-block mermaid-block-loading" aria-label="Rendering Mermaid diagram" />
+    ) : (
+      <div
+        className="mermaid-block"
+        dangerouslySetInnerHTML={{ __html: svg }}
+      />
+    );
+
+  return (
+    <div className="markdown-code-block">
+      <div className="markdown-code-header">
+        <span className="markdown-code-lang">mermaid</span>
+        {previewButton}
+      </div>
+      {body}
+    </div>
+  );
+}
+
+interface CodeBlockProps {
+  code: string;
+  lang: string;
+  headerAction?: ReactNode;
+}
+
+/**
+ * Syntax-highlighted code block with copy button.
+ * Used as the "source" view for mermaid blocks and for all non-mermaid code fences.
+ */
+export function CodeBlock({ code, lang, headerAction }: CodeBlockProps) {
+  const { isDark } = useTheme();
+  const [copied, setCopied] = useState(false);
+
+  const copy = () => {
+    copyText(code).then(() => {
+      setCopied(true);
+      setTimeout(() => setCopied(false), 1500);
+    });
+  };
+
+  return (
+    <div className="markdown-code-block">
+      <div className="markdown-code-header">
+        <span className="markdown-code-lang">{lang || "text"}</span>
+        <div className="markdown-code-actions">
+          {headerAction}
+          <button
+            onClick={copy}
+            className="markdown-code-action"
+          >
+            {copied ? "copied" : "copy"}
+          </button>
+        </div>
+      </div>
+      <SyntaxHighlighter
+        language={lang || "text"}
+        style={isDark ? vscDarkPlus : vs}
+        showLineNumbers
+        lineNumberStyle={{ color: "var(--text-dim)", fontStyle: "normal" }}
+        customStyle={{
+          margin: 0,
+          padding: "11px 13px",
+          fontSize: 12.5,
+          lineHeight: 1.62,
+          borderRadius: 0,
+          background: "color-mix(in srgb, var(--bg) 92%, var(--bg-panel))",
+        }}
+        codeTagProps={{ style: { fontFamily: "var(--font-mono)" } }}
+      >
+        {code}
+      </SyntaxHighlighter>
+    </div>
+  );
+}
